Rides endpoint: Added POST and GET endpoints
