+++ src/Models/RideSModel.js
@@ -1,11 +1,16 @@
-const mongoose = require('mongoose');
+const mongoose = require("mongoose");
 
 const RidesSchema = new mongoose.Schema({
-      user_id: String,
-      destination: String,
+  ride_id: String,
+  users: [],
+  date: Date,
+  origin: String,
+  destination: String,
+  uber: Boolean,
+  price: Number,
+  payment_type: Boolean,
 });
 
-const RidesModel = mongoose.model('RideModel', RidesSchema);
+const RidesModel = mongoose.model("RideModel", RidesSchema);
 
 module.exports = RidesModel;
-
