+++ src/routes/index.js
@@ -1,5 +1,6 @@
 const express = require('express');
 const test = require('./test.route');
+const rides= require('./rides.route');
 
 const router = express.Router();
 
@@ -8,6 +9,10 @@ const defaultRoutes = [
         path: "/test",
         route: test,
     },
+    {
+        path: "/rides",
+        route: rides,
+    }
 ];
 
 defaultRoutes.forEach((route) => {
